게시글 작성자 검사 수정 삭제

+++ src/main/kotlin/org/example/coha/domain/post/service/PostServiceImpl.kt
@@ -1,12 +1,14 @@
 package org.example.coha.domain.post.service
 
 import org.example.coha.domain.exception.ModelNotFoundException
+import org.example.coha.domain.exception.UnauthorizedAccess
 import org.example.coha.domain.post.dto.CreatePostRequest
 import org.example.coha.domain.post.dto.PostResponse
 import org.example.coha.domain.post.dto.PostWithReplyResponse
 import org.example.coha.domain.post.dto.UpdatePostRequest
 import org.example.coha.domain.post.repository.PostRepository
 import org.springframework.data.repository.findByIdOrNull
+import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
@@ -18,7 +20,9 @@ class PostServiceImpl(
 
     @Transactional
     override fun createPost(request: CreatePostRequest): PostResponse {
-        val post = postRepository.save(request.toPost())
+        val currentUser = SecurityContextHolder.getContext().authentication.name
+
+        val post = postRepository.save(request.toPost(currentUser))
         return PostResponse.toPostResponse(post)
 
     }
@@ -36,6 +40,8 @@ class PostServiceImpl(
     @Transactional
     override fun updatePost(postId: Long, request: UpdatePostRequest): PostResponse {
         val savedPost = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("Post", request.id)
+        val currentUser = SecurityContextHolder.getContext().authentication.name
+        if(savedPost.author != currentUser) throw UnauthorizedAccess()
         savedPost.content = request.content
         return PostResponse.toPostResponse(savedPost)
     }
@@ -43,6 +49,9 @@ class PostServiceImpl(
 
     @Transactional
     override fun deletePost(postId: Long) {
+        val post = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("Post", postId)
+        val currentUser = SecurityContextHolder.getContext().authentication.name
+        if(post.author != currentUser) throw UnauthorizedAccess()
         postRepository.deleteById(postId)
     }
